requirements.txt
# Core dependencies
customtkinter>=5.2.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
Pillow>=10.0.0

# Video processing
opencv-python>=4.8.0
moviepy>=1.0.3,<2.0

# Social media APIs
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
youtube-transcript-api>=0.6.2

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
nicegui>=1.4.0
pywebview>=5.0.0

# Phase 1 Backend
cryptography>=41.0.0
flask>=3.0.0
flask-cors>=4.0.0

# Phase 2 AI Engine (optional - install separately if needed)
# TTS>=0.22.0  # Coqui TTS for voice cloning (large download)

# Testing
pytest>=8.0.0

# Database
# supabase>=2.0.0  # Optional for cloud sync

# Audio
pyttsx3>=2.90

# Note: FFmpeg and Ollama must be installed separately
# FFmpeg: winget install FFmpeg
# Ollama: Download from https://ollama.ai